DataInjectLogic/requirements.txt
redis-py 
json